pyproject.toml
[project]
name = "wazuh-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.13.1",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
